Add Mesh Data Transfer community add-on created by Maurizio Memoli
Downloaded from Maurizio's Gumroad for free. Sent to me by  Dzooworks over Discord.

The code was and still is quite messy, I cleaned up imports and formatting, but there's hundreds of lines of commented out code in operators.py. Did not yet investigate why, or even test the add-on yet. But it looks like something I've wanted for a long time, and parts of it we probably implemented as part of the Asset Pipeline.